requirementsa.txt
Django==1.7.10
django-disqus==0.5
funcsigs==0.4
mock==1.3.0
oauthlib==1.0.3
pbr==1.8.0
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.12
reportlab==3.2.0
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
Unipath==1.1
virtualenv==13.1.2
virtualenvwrapper-win==1.2.1
wheel==0.24.0
